[refactor] handlers logic moved to server folder, controllers created
