+++ rust/src/controllers/mod.rs
@@ -1,0 +1,4 @@
+pub mod devices;
+pub mod emergency;
+pub mod invitations;
+pub mod telemetry;
