--- rust/src/http_gateway/mod.rs
@@ -1,2 +1,0 @@
-pub mod handlers;
-
