+++ rust/src/messaging/mod.rs
@@ -22,13 +22,20 @@ use chrono::Utc;
 use postgres::NoTls;
 use r2d2::PooledConnection;
 use r2d2_postgres::PostgresConnectionManager;
-use crate::db::devices::get_subscriber_device_ids;
-
+use crate::controllers::devices::get_subscriber_device_ids;
+use crate::lang::TranslationIds;
 use crate::model::{
+    responses::Errors::APIInternalError,
     notifications::{NotificationData, PushNotification},
     telemetry::{TelemetryUpdate, TelemetryWebsocketUpdate},
     devices::OS,
 };
+
+use std::{
+    sync::{Arc, Mutex},
+};
+use rocket::State;
+
 use crate::constants::DATE_FORMAT;
 
 pub mod slack;
@@ -47,6 +54,14 @@ pub fn get_rabbitmq_uri() -> String {
         "amqp://{}:{}@{}/{}",
         rabbitmq_user, rabbitmq_pass, rabbitmq_host, rabbitmq_vhost
     )
+}
+
+pub fn unlock_channel(state: State<Arc<Mutex<Connection>>>) -> Result<Channel, APIInternalError> {
+    let mut guard = state.lock().expect("Rabbit Connection was poisoned");
+    guard.open_channel(None).map_err(|w| APIInternalError {
+        msg: TranslationIds::BackendIssue,
+        engineering_error: Some(w.to_string()),
+    })
 }
 
 pub fn send_ws_message(
